+++ app/Http/Controllers/Budgeting/Referensi/programControllerAdum.php
@@ -15,6 +15,7 @@ use App\Model\Impact;
 use View;
 use Carbon;
 use Response;
+use Auth;
 use DB;
 use Illuminate\Support\Facades\Input;
 class programControllerAdum extends Controller
@@ -45,7 +46,12 @@ class programControllerAdum extends Controller
         $aksi           = '';
     	$view 			= array();
     	foreach ($data as $data) {
+            if(Auth::user()->level==8){
     		$aksi 		= '<div class="action visible pull-right"><a onclick="return showCapaian(\''.$data->PROGRAM_KODE.'\')" class="action-edit"><i class="mi-eye"></i></a><a onclick="return ubahProgram(\''.$data->PROGRAM_KODE.'\')" class="action-edit"><i class="mi-edit"></i></a><a onclick="return hapusProgram(\''.$data->PROGRAM_KODE.'\')" class="action-delete"><i class="mi-trash"></i></a></div>';
+            else{
+                $aksi       = '<div class="action visible pull-right"><a onclick="return showCapaian(\''.$data->PROGRAM_KODE.'\')" class="action-edit"><i class="mi-eye"></i></a></div>';
+            }
+
     		array_push($view, array( 'id_porgram'		=>$data->PROGRAM_KODE,
                                      'OPSI'				=>$aksi,
                                      'PROGRAM'			=>"<i class='mi-caret-down m-r-sm'></i>".$data->PROGRAM_KODE." - ".$data->PROGRAM_NAMA));
@@ -63,7 +69,11 @@ class programControllerAdum extends Controller
     	$view 			= array();
     	foreach ($data as $data) {
             // <a onclick="return showIndikatorGiat(\''.$data->KEGIATAN_KODE.'\')" class="action-edit"><i class="mi-eye"></i></a>
-    		$aksi 		= '<div class="action visible pull-right"><a onclick="return ubahGiat(\''.$data->KEGIATAN_KODE.'\')" class="action-edit"><i class="mi-edit"></i></a><a onclick="return hapusGiat(\''.$data->KEGIATAN_KODE.'\')" class="action-delete"><i class="mi-trash"></i></a></div>';
+            if(Auth::user()->level==8){
+    		  $aksi 		= '<div class="action visible pull-right"><a onclick="return ubahGiat(\''.$data->KEGIATAN_KODE.'\')" class="action-edit"><i class="mi-edit"></i></a><a onclick="return hapusGiat(\''.$data->KEGIATAN_KODE.'\')" class="action-delete"><i class="mi-trash"></i></a></div>';
+            }
+            else $aksi = '';
+
     		array_push($view, array( 'KEGIATAN_KODE'  	=>$data->KEGIATAN_KODE,
                                      'KEGIATAN_NAMA'	=>$data->KEGIATAN_NAMA,
                                      'AKSI'				=>$aksi));
